+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [0.27.0] - 2021-01-15
+### Changed
+ - plugin system changed to require a factory function instead of a singleton object ([#272](https://github.com/rokucommunity/brighterscript/pull/272))
+
+
+
 ## [0.26.0] - 2021-01-14
 ### Added
  - proper XML AST support
@@ -820,3 +826,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.24.2]:   https://github.com/rokucommunity/brighterscript/compare/v0.24.1...v0.24.2
 [0.25.0]:   https://github.com/rokucommunity/brighterscript/compare/v0.24.2...v0.25.0
 [0.26.0]:   https://github.com/rokucommunity/brighterscript/compare/v0.25.0...v0.26.0
+[0.27.0]:   https://github.com/rokucommunity/brighterscript/compare/v0.26.0...v0.27.0